refactor: リファクタリング
カメラと音源とプレイヤーのCSに接頭辞追加
カメラのCSにコメント追加とアクセス指定子見直し
音源のCSに不必要なコメントの削除
