+++ Assets/Programmer/ScriptableObject/BackGround_BGM_DataSO.asset
@@ -9,9 +9,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7a03b83d4f157d843b3fb01c10e0a104, type: 3}
+  m_Script: {fileID: 0}
   m_Name: BackGround_BGM_DataSO
-  m_EditorClassIdentifier: Assembly-CSharp::BGMDataBase
+  m_EditorClassIdentifier: Assembly-CSharp::BackGround_BGM_DataBase
   bgmDatas:
   - sceneName: 0
     cueName: 0
