--- Assets/Programmer/Scripts/Audio/PlaySE_3D.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7c3678cc0d9d456448a215970fa53130
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
