+++ Assets/Programmer/Scripts/Audio/SO_BackGround_BGM_DataBase.cs
@@ -4,7 +4,6 @@
  *    元浪梨緒
  * ----------------------------------------------------------
  * 2026-04-22 | 初回作成
- * 
  */
 
 using System;
